refactor(database): split RTDB CDD replay tests
