+++ packages/conformance/registry/rtdb.ts
@@ -1043,7 +1043,7 @@ export const rtdbRegistry = {
           riskReasons: ["asserts listener semantics"],
           automation: "oracle-backed",
           oracleObservations: ["rtdb-modular-priority-contract"],
-          conformanceTests: ["packages/pyric/test/database/modular/sandbox-child-events.test.ts","packages/pyric/test/database/modular/cdd-climb-cases.ts"],
+          conformanceTests: ["packages/pyric/test/database/modular/sandbox-child-events.test.ts","packages/pyric/test/database/modular/priority-metadata-cdd.test.ts"],
           rowNumber: 46,
         }),
         row5({
@@ -1257,10 +1257,10 @@ export const rtdbRegistry = {
           featureKeys: ["runTransaction"],
           behavior: "**Divergence:** two ordinary concurrent `runTransaction` calls are serialized by the in-process backend, so their update functions are not retried with Firebase's captured contention counts. A synchronous re-entrant conflicting write does trigger a deterministic retry.",
           status: "diverged-documented",
-          evidence: "Oracle `rtdb-modular-concurrent-transforms` captures invocation counts [2, 3]; `unit:modular/cdd-climb-cases.ts` pins the sandbox's [1, 1] ordinary-concurrency boundary and separately covers its deterministic re-entrant retry seam.",
+          evidence: "Oracle `rtdb-modular-concurrent-transforms` captures invocation counts [2, 3]; `unit:modular/transaction-contention-cdd.test.ts` pins the sandbox's [1, 1] ordinary-concurrency boundary and separately covers its deterministic re-entrant retry seam.",
           automation: "oracle-backed",
           oracleObservations: ["rtdb-modular-concurrent-transforms"],
-          conformanceTests: ["packages/pyric/test/database/modular/cdd-climb-cases.ts"],
+          conformanceTests: ["packages/pyric/test/database/modular/transaction-contention-cdd.test.ts"],
           conformanceDisposition: "pending-fix",
           rowNumber: 37,
         }),
@@ -1637,13 +1637,13 @@ export const rtdbRegistry = {
           featureKeys: ["onChildAdded","onChildChanged","onChildRemoved","onChildMoved"],
           behavior: "Child-listener callbacks receive Firebase's `previousChildName` second argument, including initial replay and ordered-query movement.",
           status: "conforms",
-          evidence: "Oracle `rtdb-modular-child-previous-name` captures initial replay plus add/change/remove/move predecessor values; replayed by `unit:modular/cdd-climb-cases.ts`.",
+          evidence: "Oracle `rtdb-modular-child-previous-name` captures initial replay plus add/change/remove/move predecessor values; replayed by `unit:modular/listener-lifecycle-cdd.test.ts`.",
           risk: ["specific-value","listener"],
           riskScore: 4,
           riskReasons: ["asserts 1 specific value(s)","asserts listener semantics"],
           automation: "oracle-backed",
           oracleObservations: ["rtdb-modular-child-previous-name"],
-          conformanceTests: ["packages/pyric/test/database/modular/cdd-climb-cases.ts"],
+          conformanceTests: ["packages/pyric/test/database/modular/listener-lifecycle-cdd.test.ts"],
           conformanceChecks: [{
             finding: "RTDB-M75",
             observation: "rtdb-modular-child-previous-name",
@@ -1659,13 +1659,13 @@ export const rtdbRegistry = {
           featureKeys: ["onValue","onChildAdded","onChildChanged","onChildRemoved","onChildMoved"],
           behavior: "Value and child listener overloads accept a cancellation callback and deliver Firebase-shaped permission errors when a listen is denied or revoked; a callbackless listener that loses auth access is terminal and does not resurrect after sign-in.",
           status: "conforms",
-          evidence: "Oracle `rtdb-modular-listener-cancellation` captures all five registrars returning normally, asynchronous initial-denial cancellation, one revocation callback each, and the Error/PERMISSION_DENIED/message shape; replayed in `unit:modular/cdd-climb-cases.ts`.",
+          evidence: "Oracle `rtdb-modular-listener-cancellation` captures all five registrars returning normally, asynchronous initial-denial cancellation, one revocation callback each, and the Error/PERMISSION_DENIED/message shape; replayed in `unit:modular/listener-lifecycle-cdd.test.ts`.",
           risk: ["listener","rules-denial"],
           riskScore: 4,
           riskReasons: ["asserts listener semantics","asserts rules-denial behavior"],
           automation: "oracle-backed",
           oracleObservations: ["rtdb-modular-listener-cancellation"],
-          conformanceTests: ["packages/pyric/test/database/modular/cdd-climb-cases.ts"],
+          conformanceTests: ["packages/pyric/test/database/modular/listener-lifecycle-cdd.test.ts"],
           conformanceChecks: [{
             finding: "RTDB-M75A",
             observation: "rtdb-modular-listener-cancellation",
@@ -1680,12 +1680,12 @@ export const rtdbRegistry = {
           featureKeys: ["onChildAdded","onChildChanged","onChildRemoved","onChildMoved","query"],
           behavior: "All child-listener functions accept `Query` inputs; add/change/remove events respect the active query window.",
           status: "conforms",
-          evidence: "`unit:modular/cdd-climb-cases.ts` and `unit:modular/sandbox-child-events.test.ts` cover initial windows plus entering, leaving, and changing members.",
+          evidence: "`unit:modular/listener-lifecycle-cdd.test.ts` and `unit:modular/sandbox-child-events.test.ts` cover initial windows plus entering, leaving, and changing members.",
           risk: ["listener"],
           riskScore: 2,
           riskReasons: ["asserts listener semantics"],
           automation: "unit-backed",
-          conformanceTests: ["packages/pyric/test/database/modular/cdd-climb-cases.ts","packages/pyric/test/database/modular/sandbox-child-events.test.ts"],
+          conformanceTests: ["packages/pyric/test/database/modular/listener-lifecycle-cdd.test.ts","packages/pyric/test/database/modular/sandbox-child-events.test.ts"],
         }),
         row5({
           rowRef: "M75c",
@@ -1699,7 +1699,7 @@ export const rtdbRegistry = {
           riskReasons: ["asserts listener semantics","asserts a specific listener fire count","asserts 1 specific value(s)"],
           automation: "oracle-backed",
           oracleObservations: ["rtdb-modular-child-previous-name","rtdb-modular-childchanged-cofire-with-childmoved","rtdb-modular-onchildmoved-previouschildname-sequencing"],
-          conformanceTests: ["packages/pyric/test/database/modular/cdd-climb-cases.ts","packages/pyric/test/database/modular/oracle-conformance.test.ts"],
+          conformanceTests: ["packages/pyric/test/database/modular/listener-lifecycle-cdd.test.ts","packages/pyric/test/database/modular/oracle-conformance.test.ts"],
           conformanceChecks: [{
             finding: "RTDB-M75C",
             observation: "rtdb-modular-child-previous-name",
@@ -1726,13 +1726,13 @@ export const rtdbRegistry = {
           featureKeys: ["onChildAdded","onChildChanged","onChildRemoved","onChildMoved"],
           behavior: "Child listener overloads accept `ListenOptions` directly or after a cancellation callback; `onlyOnce` stops delivery after Firebase's captured initial/event batch.",
           status: "conforms",
-          evidence: "Oracle `rtdb-modular-child-listener-only-once` captures all four listener families, both overload shapes, the existing-child replay batch, and one-shot post-registration delivery; replayed by `unit:modular/cdd-climb-cases.ts`.",
+          evidence: "Oracle `rtdb-modular-child-listener-only-once` captures all four listener families, both overload shapes, the existing-child replay batch, and one-shot post-registration delivery; replayed by `unit:modular/listener-lifecycle-cdd.test.ts`.",
           risk: ["listener","listener-fire-count"],
           riskScore: 4,
           riskReasons: ["asserts listener semantics","asserts a specific listener fire count"],
           automation: "oracle-backed",
           oracleObservations: ["rtdb-modular-child-listener-only-once"],
-          conformanceTests: ["packages/pyric/test/database/modular/cdd-climb-cases.ts"],
+          conformanceTests: ["packages/pyric/test/database/modular/listener-lifecycle-cdd.test.ts"],
           conformanceChecks: [{
             finding: "RTDB-M75D",
             observation: "rtdb-modular-child-listener-only-once",
@@ -1953,13 +1953,13 @@ export const rtdbRegistry = {
           featureKeys: ["setPriority","setWithPriority"],
           behavior: "`setPriority` and `setWithPriority` store, replace, preserve, and clear valid RTDB priority metadata; parent `exportVal()` and `toJSON()` recursively include descendant priority metadata.",
           status: "conforms",
-          evidence: "Oracle `rtdb-modular-priority-contract` captures round-trip/export shape, update/transaction preservation, plain-set clearing, and explicit clearing; replayed by `unit:modular/cdd-climb-cases.ts`.",
+          evidence: "Oracle `rtdb-modular-priority-contract` captures round-trip/export shape, update/transaction preservation, plain-set clearing, and explicit clearing; replayed by `unit:modular/priority-metadata-cdd.test.ts`.",
           risk: ["specific-field"],
           riskScore: 1,
           riskReasons: ["asserts a specific field/property value"],
           automation: "oracle-backed",
           oracleObservations: ["rtdb-modular-priority-contract"],
-          conformanceTests: ["packages/pyric/test/database/modular/cdd-climb-cases.ts"],
+          conformanceTests: ["packages/pyric/test/database/modular/priority-metadata-cdd.test.ts"],
           conformanceChecks: [{
             finding: "RTDB-M89",
             observation: "rtdb-modular-priority-contract",
@@ -1975,13 +1975,13 @@ export const rtdbRegistry = {
           featureKeys: ["orderByPriority"],
           behavior: "`orderByPriority` orders by Firebase priority with key tie-breaking and composes with bounds, equality, and limits; plain snapshots and unconstrained queries use the same default priority index, and invalid priority bounds throw synchronously.",
           status: "conforms",
-          evidence: "Oracle `rtdb-modular-priority-contract` captures priority order, key tie-breaking, bounds, and limits; replayed by `unit:modular/cdd-climb-cases.ts`.",
+          evidence: "Oracle `rtdb-modular-priority-contract` captures priority order, key tie-breaking, bounds, and limits; replayed by `unit:modular/priority-metadata-cdd.test.ts`.",
           risk: ["specific-value"],
           riskScore: 1,
           riskReasons: ["asserts 1 specific value(s)"],
           automation: "oracle-backed",
           oracleObservations: ["rtdb-modular-priority-contract"],
-          conformanceTests: ["packages/pyric/test/database/modular/cdd-climb-cases.ts"],
+          conformanceTests: ["packages/pyric/test/database/modular/priority-metadata-cdd.test.ts"],
           conformanceChecks: [{
             finding: "RTDB-M90",
             observation: "rtdb-modular-priority-contract",
@@ -1997,13 +1997,13 @@ export const rtdbRegistry = {
           featureKeys: ["setPriority","onChildMoved","runTransaction"],
           behavior: "Priority changes participate in plain and explicitly ordered child movement/listener sequencing, transactions, and updates without losing metadata.",
           status: "conforms",
-          evidence: "Oracle `rtdb-modular-priority-contract` captures priority movement, update/transaction preservation, and descendant delivery when ancestor replacement clears metadata; replayed by `unit:modular/cdd-climb-cases.ts` and `unit:modular/priority-listeners.test.ts`. Disconnect priority execution is owned separately by M83.",
+          evidence: "Oracle `rtdb-modular-priority-contract` captures priority movement, update/transaction preservation, and descendant delivery when ancestor replacement clears metadata; replayed by `unit:modular/priority-metadata-cdd.test.ts` and `unit:modular/priority-listeners.test.ts`. Disconnect priority execution is owned separately by M83.",
           risk: ["listener","atomicity","specific-field"],
           riskScore: 5,
           riskReasons: ["asserts listener semantics","asserts transaction/batch atomicity","asserts a specific field/property value"],
           automation: "oracle-backed",
           oracleObservations: ["rtdb-modular-priority-contract"],
-          conformanceTests: ["packages/pyric/test/database/modular/cdd-climb-cases.ts","packages/pyric/test/database/modular/priority-listeners.test.ts"],
+          conformanceTests: ["packages/pyric/test/database/modular/priority-metadata-cdd.test.ts","packages/pyric/test/database/modular/priority-listeners.test.ts"],
           conformanceChecks: [{
             finding: "RTDB-M91",
             observation: "rtdb-modular-priority-contract",
@@ -2090,11 +2090,11 @@ export const rtdbRegistry = {
           surface: "rtdb-modular",
           behavior: "`getDatabase(ctx)` returns a tagged sandbox-target handle (frozen identity)",
           status: "conforms",
-          evidence: "`unit:modular/cdd-climb-cases.ts` asserts the returned Database carries a sandbox target.",
+          evidence: "`unit:modular/database-instances-cdd.test.ts` asserts the returned Database carries a sandbox target.",
           risk: ["structural"],
           riskReasons: ["structural / routing-only claim"],
           automation: "unit-backed",
-          conformanceTests: ["packages/pyric/test/database/modular/cdd-climb-cases.ts"],
+          conformanceTests: ["packages/pyric/test/database/modular/database-instances-cdd.test.ts"],
           aliases: ["rtdb#94"],
         }),
         row6({
@@ -2102,11 +2102,11 @@ export const rtdbRegistry = {
           surface: "rtdb-modular",
           behavior: "`getDatabase(sandbox)` returns a tagged sandbox-live handle (per-op identity)",
           status: "conforms",
-          evidence: "`unit:modular/cdd-climb-cases.ts` asserts the returned Database carries a sandbox-live target.",
+          evidence: "`unit:modular/database-instances-cdd.test.ts` asserts the returned Database carries a sandbox-live target.",
           risk: ["structural"],
           riskReasons: ["structural / routing-only claim"],
           automation: "sandbox-only",
-          conformanceTests: ["packages/pyric/test/database/modular/cdd-climb-cases.ts"],
+          conformanceTests: ["packages/pyric/test/database/modular/database-instances-cdd.test.ts"],
           exceptionReason: "sandbox-only behavior — no prod equivalent to observe",
           aliases: ["rtdb#95"],
         }),
@@ -2115,12 +2115,12 @@ export const rtdbRegistry = {
           surface: "rtdb-modular",
           behavior: "Inactive canonical `firebase/database` imports remain the upstream package; the mirror does not create tagged production targets",
           status: "conforms",
-          evidence: "Package-resolution boundary assertion in `unit:modular/cdd-climb-cases.ts` proves a real Firebase Database remains untagged and is rejected by the mirror.",
+          evidence: "Package-resolution boundary assertion in `unit:modular/database-instances-cdd.test.ts` proves a real Firebase Database remains untagged and is rejected by the mirror.",
           risk: ["specific-value"],
           riskScore: 2,
           riskReasons: ["asserts 1 specific value(s)"],
           automation: "unit-backed",
-          conformanceTests: ["packages/pyric/test/database/modular/cdd-climb-cases.ts"],
+          conformanceTests: ["packages/pyric/test/database/modular/database-instances-cdd.test.ts"],
           aliases: ["rtdb#96"],
         }),
         row6({
@@ -2142,9 +2142,9 @@ export const rtdbRegistry = {
           surface: "rtdb-modular",
           behavior: "Two `getDatabase(sandbox)` calls share state (same underlying `LocalEnvironment`)",
           status: "conforms",
-          evidence: "`unit:modular/cdd-climb-cases.ts` writes through one handle and reads through another handle owned by the same sandbox.",
+          evidence: "`unit:modular/database-instances-cdd.test.ts` writes through one handle and reads through another handle owned by the same sandbox.",
           automation: "unit-backed",
-          conformanceTests: ["packages/pyric/test/database/modular/cdd-climb-cases.ts"],
+          conformanceTests: ["packages/pyric/test/database/modular/database-instances-cdd.test.ts"],
           aliases: ["rtdb#98"],
         }),
         row6({
@@ -2152,11 +2152,11 @@ export const rtdbRegistry = {
           surface: "rtdb-modular",
           behavior: "Handle dispatch by `TARGET_SYMBOL` brand — refs route to their owning target via a `refToTarget` WeakMap (mirror of firestore's pattern)",
           status: "conforms",
-          evidence: "`unit:modular/cdd-climb-cases.ts` proves references from independent sandbox targets route to independent backends.",
+          evidence: "`unit:modular/database-instances-cdd.test.ts` proves references from independent sandbox targets route to independent backends.",
           risk: ["structural"],
           riskReasons: ["structural / routing-only claim"],
           automation: "unit-backed",
-          conformanceTests: ["packages/pyric/test/database/modular/cdd-climb-cases.ts"],
+          conformanceTests: ["packages/pyric/test/database/modular/database-instances-cdd.test.ts"],
           aliases: ["rtdb#99"],
         }),
       ],
@@ -2582,12 +2582,12 @@ export const rtdbRegistry = {
           surface: "rtdb-modular",
           behavior: "The returned value from `onValue(ref, cb)` is the unsubscribe function (NOT an object); calling it removes the listener",
           status: "conforms",
-          evidence: "`unit:modular/cdd-climb-cases.ts` asserts the return is a function and that invoking it stops subsequent delivery; production behavior is adjacent to oracle-backed row #131.",
+          evidence: "`unit:modular/listener-lifecycle-cdd.test.ts` asserts the return is a function and that invoking it stops subsequent delivery; production behavior is adjacent to oracle-backed row #131.",
           risk: ["listener"],
           riskScore: 2,
           riskReasons: ["asserts listener semantics"],
           automation: "unit-backed",
-          conformanceTests: ["packages/pyric/test/database/modular/cdd-climb-cases.ts"],
+          conformanceTests: ["packages/pyric/test/database/modular/listener-lifecycle-cdd.test.ts"],
           aliases: ["rtdb#132"],
         }),
       ],
@@ -2958,13 +2958,13 @@ export const rtdbRegistry = {
           featureKeys: ["increment"],
           behavior: "Two concurrent `increment` calls interleave correctly (last-write-wins is NOT the contract — both deltas accumulate)",
           status: "conforms",
-          evidence: "Oracle `rtdb-modular-concurrent-transforms` uses two independent clients and captures terminal 5 after concurrent +2/+3 increments; replayed by `unit:modular/cdd-climb-cases.ts`.",
+          evidence: "Oracle `rtdb-modular-concurrent-transforms` uses two independent clients and captures terminal 5 after concurrent +2/+3 increments; replayed by `unit:modular/transaction-contention-cdd.test.ts`.",
           risk: ["listener","sentinel"],
           riskScore: 4,
           riskReasons: ["asserts listener semantics","asserts sentinel or atomic transform semantics"],
           automation: "oracle-backed",
           oracleObservations: ["rtdb-modular-concurrent-transforms"],
-          conformanceTests: ["packages/pyric/test/database/modular/cdd-climb-cases.ts"],
+          conformanceTests: ["packages/pyric/test/database/modular/transaction-contention-cdd.test.ts"],
           conformanceChecks: [{
             finding: "RTDB-157",
             observation: "rtdb-modular-concurrent-transforms",
@@ -3026,13 +3026,13 @@ export const rtdbRegistry = {
           surface: "rtdb-modular",
           behavior: "Concurrent contention — if another client writes between the read and write, the update fn is retried with the new current value (typically up to 25 retries by default)",
           status: "diverged-documented",
-          evidence: "Oracle `rtdb-modular-concurrent-transforms` captures callback counts [2, 3] across two clients; the synchronous in-process backend serializes the same ordinary calls with [1, 1], pinned by `unit:modular/cdd-climb-cases.ts`.",
+          evidence: "Oracle `rtdb-modular-concurrent-transforms` captures callback counts [2, 3] across two clients; the synchronous in-process backend serializes the same ordinary calls with [1, 1], pinned by `unit:modular/transaction-contention-cdd.test.ts`.",
           risk: ["listener"],
           riskScore: 2,
           riskReasons: ["asserts listener semantics"],
           automation: "oracle-backed",
           oracleObservations: ["rtdb-modular-concurrent-transforms"],
-          conformanceTests: ["packages/pyric/test/database/modular/cdd-climb-cases.ts"],
+          conformanceTests: ["packages/pyric/test/database/modular/transaction-contention-cdd.test.ts"],
           conformanceDisposition: "pending-fix",
           aliases: ["rtdb#161"],
         }),
